Add some package structure
