+++ src/pymmcore_openscan/__init__.py
@@ -1,10 +1,12 @@
-"""A set of widgets for OpenScan, built atop the pymmcore-plus module."""
+"""A set of widgets for OpenScan, built atop pymmcore-plus, for pymmcore-gui."""
 
 from importlib.metadata import PackageNotFoundError, version
+
+from pymmcore_openscan._util import augment_pymmcore_gui
 
 try:
     __version__ = version("pymmcore-openscan")
 except PackageNotFoundError:
     __version__ = "uninstalled"
 
-__all__: list[str] = []
+__all__: list[str] = ["__version__", "augment_pymmcore_gui"]
