+++ src/pymmcore_openscan/widgets/__init__.py
@@ -1,0 +1,12 @@
+"""A set of widgets for OpenScan, built atop the pymmcore-plus module."""
+
+from pymmcore_openscan.widgets.bh_dcc_dcu import DCCWidget, DCUWidget
+from pymmcore_openscan.widgets.image_collection_params import ImageCollectionParameters
+from pymmcore_openscan.widgets.spc import SPCRateCounters
+
+__all__: list[str] = [
+    "DCCWidget",
+    "DCUWidget",
+    "ImageCollectionParameters",
+    "SPCRateCounters",
+]
