+++ src/pymmcore_openscan/widgets/image_collection_params.py
